+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # 0.1.1
+* Rename Mock to Capture
+* Add (*=>) combinator